requirements.txt
llama-index-core>=0.10.0
gradient>=3.8.0
python-dotenv>=1.0.0